+++ example_logging.py
@@ -44,6 +44,8 @@ logging.basicConfig(level=logging.DEBUG,
                     style="{")
 
 
+# To use this library you need a valid UNICORN Binance Suite License:
+# https://medium.lucit.tech/-87b0088124a8
 ubwa = unicorn_binance_websocket_api.BinanceWebSocketApiManager(exchange="binance.com",
                                                                 output_default="UnicornFy")