+++ reset.sh
@@ -1,0 +1,12 @@
+#!/usr/bin/env bash
+
+rm *.py.log
+
+rm build -r
+rm dist -r
+rm *.egg-info -r
+
+rm unicorn_fy/*.c
+rm unicorn_fy/*.html
+rm unicorn_fy/*.dll
+rm unicorn_fy/*.so