+++ sphinx/install_sphinx.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 python3 -m pip install sphinx --upgrade
 python3 -m pip install python-docs-theme-lucit --upgrade
-python3 -m pip install recommonmark --upgrade
+python3 -m pip install rich --upgrade
+python3 -m pip install myst-parser --upgrade
 python3 -m pip install sphinx-markdown-tables --upgrade