+++ sphinx/release_docs.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env bash
+#!/usr/bin/bash
 
 cp ./build/html/* ../docs -R
 rm ./build/html/ -r