+++ sphinx/source/conf.py
@@ -22,7 +22,7 @@ copyright = '2019-2023, LUCIT Systems and Development and Oliver Zehentleitner'
 author = 'LUCIT Systems and Development'
 
 # The full version, including alpha/beta/rc tags
-release = '0.12.2'
+release = '0.13.0'
 
 html_last_updated_fmt = "%b %d %Y at %H:%M (CET)"